Remove local AudioSocketHost, and use common providerlib implementation
Also removes dependency for a short[]-only transmission, allowing a fully
byte[] pipeline to reduce/remove the necessity of doing byte<=>short
conversion (NativePlayer takes bytes anyway, and sockets as well)

+++ src/main/java/org/omnirom/music/service/AudioSink.java
@@ -25,7 +25,7 @@ public interface AudioSink {
      * @param numframes Number of frames
      * @return The number of bytes actually written
      */
-    public int write(short[] frames, int numframes);
+    public int write(byte[] frames, int numframes);
 
     /**
      * Returns the number of samples written since the last call to flushSamples() has been made
